+++ VRGame/Assets/Tests/OccipitalLobeGame/ObjectMatchGame/PlayMode/GuessBoxPlayModeTests.cs
@@ -225,5 +225,8 @@ public class GuessBoxPlayModeTests
         public string GetCurrentGuessID() => Controller?.GetCurrentGuessID() ?? "";
         public void RemovePotentialGuess() => Controller?.RemovePotentialGuess();
         public void SubmitGuess() => Controller?.SubmitGuess();
+        public void ExitLevel() => Controller?.ExitLevel();
+        public void LeaveTutorial() => Controller?.LeaveTutorial();
+        public void Update() => Controller?.Update();
     }
 }